--- scripts/bank update.bat
@@ -1,1 +1,0 @@
-"d:\Home\env\Scripts\python.exe" "D:\Home\scripts\accounts.py"
